Move horizontal transform benchmarks to package
